requirements.txt
# langgraph
# langsmith
# langchain_anthropic
# python-dotenv
# langchain
# langchain-community>=0.2.11
# tavily-python
# requests
# langchain-openai
# langchainhub
# aiohttp
# PyPDF2
# reportlab
# fastapi
# pydantic
# uvicorn
# gunicorn
# python-multipart
# pypdf
# langchain-unstructured
# unstructured-client
# unstructured
# unstructured[pdf]
# python-magic
# PyMuPDF
# pillow 
# python-docx 
# docx2pdf
# pandas
# antiword


langgraph==0.2.62
langsmith==0.2.10
langchain_anthropic==0.3.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
tavily-python==0.5.0
requests==2.32.3
langchain-openai==0.3.0
langchainhub==0.1.21
aiohttp==3.11.11
PyPDF2==3.0.1
reportlab==4.2.5
fastapi==0.115.6
pydantic==2.9.2
uvicorn==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20
pypdf==5.1.0
python-magic==0.4.27
PyMuPDF==1.25.1
pillow==11.1.0
python-docx==1.1.2
docx2pdf==0.1.8
pandas==2.2.3
antiword==0.1.0
openai==1.59.6
python-dateutil==2.9.0.post0
llama-cloud-services==0.6.62
docling==2.47.1